Update mojoal.c
This is an update Ryan Sent to the group as supposed fix for the atomicLock .
When I implement it in lime* and use it , there is no sound coming out. But the permanent lock (stack in a loop) doesn't happen anymore. 
And all sounds are detected properly, even the event when they finish playing.